MRMScoring nach OpenMS
