+++ src/openms/source/ANALYSIS/OPENSWATH/IonMobilityScoring.cpp
@@ -39,7 +39,7 @@
 
 // scoring
 #include <OpenMS/OPENSWATHALGO/ALGO/Scoring.h>
-#include <OpenMS/OPENSWATHALGO/ALGO/MRMScoring.h>
+#include <OpenMS/ANALYSIS/OPENSWATH/MRMScoring.h>
 #include <OpenMS/ANALYSIS/OPENSWATH/SONARScoring.h>
 
 // auxiliary
