+++ src/openms/source/ANALYSIS/OPENSWATH/MasstraceCorrelator.cpp
@@ -34,7 +34,7 @@
 
 #include <OpenMS/ANALYSIS/OPENSWATH/MasstraceCorrelator.h>
 
-#include <OpenMS/OPENSWATHALGO/ALGO/MRMScoring.h>
+#include <OpenMS/ANALYSIS/OPENSWATH/MRMScoring.h>
 #include <OpenMS/OPENSWATHALGO/ALGO/Scoring.h>
 #include <OpenMS/FILTERING/NOISEESTIMATION/SignalToNoiseEstimatorMedian.h>
 #include <OpenMS/MATH/STATISTICS/StatisticFunctions.h>
