+++ src/openswathalgo/include/OpenMS/OPENSWATHALGO/ALGO/sources.cmake
@@ -4,7 +4,7 @@ set(directory include/OpenMS/OPENSWATHALGO/ALGO)
 ### list all header files of the directory here
 set(sources_list_h
 StatsHelpers.h
-MRMScoring.h
+
 Scoring.h
 )
