+++ src/tests/class_tests/openswathalgo/CMakeLists.txt
@@ -42,7 +42,7 @@ include_directories(SYSTEM ${OpenMS_INCLUDE_DIRECTORIES})
 # --------------------------------------------------------------------------
 # list of available tests
 set(openswath_algo_tests
-  MRMScoring_test
+
   Scoring_test
   Datastructures_test
   TestConvert
